backend/requirements.txt
Django>=5.2
djangorestframework
djangorestframework-simplejwt
channels
channels_redis
dj-database-url
python-dotenv
psycopg2-binary
